riscv-he: Adjust interrupt handling
1. QEMU (at least the version I use) does not clear a pending
   external interrupt when claiming from PLIC, so complete_claim
   is called right after claim to avoid spurious interrupts. The
   source of the interrupt is not disabled, and thus enabling
   the source is not necessary when ACK.

2. sbi_clear_ipi is called to clear the pending software interrupt
   bit, after the call to ipi_clear_irq.

The changes are based on QEMU with v0.6 hypervisor
extension. The changes could be wrong on real boards.

An issue is that a pending reschedule IPI can be overwritten
by anotherr eschedule IPI. This happens when a core grabs the kernel
lock and sends the second reschedule IPI right after it just sent
the first one, and before the core, which is the target of the
reschedule IPIs, has a chance to grab the lock and handles the first
reschedule IPI. The issue does not seem cause any problems at the
moment. Anyway, the code to address the issue is available. Another
commit will be made if needed.

Signed-off-by: Yanyan Shen <yshen@hybridkernel.com>