riscv-he: hack for 16KiB alignment requirement for stage-2 root page table
This is a sneaky hack that, together with an accompanying hack in Microkit,
allow us to boot a Linux VM where seL4 is a hypervisor without having
to change seL4 to support the extra alignment properly. This hack is
supposed to be temporary and the real solution is to change seL4 to have
different page table sizes for hypervisor support to adhere to the 16KiB
alignment instead of right now which is 4KiB alignment.