riscv-he: attempt to fix guest FPU access
Previously, if a guest was to access the FPU `vsstatus.FS` would not be
set which would result in an illegal instruction fault that is
delivered to the guest itself, it does *not* trap into the hypervisor.

This is consistent with S-mode software however. If seL4 in S-mode was
to cause an illegal instruction, it receives the fault.

But, this fault being transparent to the hypervisor means is at odds
with seL4's lazy FPU switching.

RFC-18 for seL4 is not yet merged at the time of this commit, but should
mean that this problem goes away for us. Until then, this patch means
that the FPU is enabled unconditionally for threads that have a vCPU
associated with them when seL4 acts as a hypervisor on RISC-V.

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>