riscv-he: fix passing faulting address for guest page faults
The motivation for this is due to unaligned-access where htval is not
sufficient to emulate. The best explanation is probably from the
specification of htval:

> If the least-significant two bits of a faulting guest physical address
> are needed, these bits are ordinarily the same as the
> least-significant two bits of the faulting virtual address in stval.
> For faults due to implicit memory accesses for VS-stage address
> translation, the least-significant two bits are instead zeros. These
> cases can be distinguished using the value provided in register htinst.

So, we read stval and then take the two least-significant bits and add
them to the address that is passed up to the user-level fault handler.

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>